+++ framework/core/src/mappers/index.ts
@@ -1,7 +1,5 @@
 export * from './models/requestDtoMapper.model';
 export * from './models/responseDtoMapper.model';
 export * from './serviceSchemaResolver';
-export * from './transformIntoInternalDtoMapper';
 export * from './types/mappers.types';
-export * from './types/internalDtoMapper.types';
 export * from './types/serviceSchemaResolver.types';